add witness schedule unit test
