+++ src/main/java/org/tron/core/db/DynamicPropertiesStore.java
@@ -14,7 +14,7 @@ import org.tron.core.config.args.Args;
 @Slf4j
 public class DynamicPropertiesStore extends TronStoreWithRevoking<BytesCapsule> {
 
-  private static final long MAINTENANCE_TIME_INTERVAL = 24 * 3600 * 1000;// (ms)
+  private static final long MAINTENANCE_TIME_INTERVAL = 300 * 1000;// (ms)
   private static final long MAINTENANCE_SKIP_SLOTS = 2;
 
   private static final byte[] LATEST_BLOCK_HEADER_TIMESTAMP = "latest_block_header_timestamp"
@@ -94,7 +94,7 @@ public class DynamicPropertiesStore extends TronStoreWithRevoking<BytesCapsule> 
 
   private static DynamicPropertiesStore instance;
 
-  public void destroy() {
+  public static void destroy() {
     instance = null;
   }
