+++ src/main/resources/config.conf
@@ -1,7 +1,8 @@
 net {
-  # type = mainnet
+//  type = mainnet
   type = testnet
 }
+
 
 storage {
   # Directory for storing persistent data
@@ -16,10 +17,8 @@ node.discovery = {
 }
 
 node {
-  # trust node for solidity node
-  trustNode = "47.254.16.55:50051"
-
-  listen.port = 18888
+  trustNode = ""
+  listen.port = 9001
 
   connection.timeout = 2
 
@@ -31,23 +30,12 @@ node {
     #    port = 30303
     #    nodeId = e437a4836b77ad9d9ffe73ee782ef2614e6d8370fcf62191a6e488276e23717147073a7ce0b444d485fff5a0c34c4577251a7a990cf80d8542e21b95aa8c5e6c
     # }
-    "47.254.16.55:18888",
-    "47.254.18.49:18888",
-    "18.188.111.53:18888",
-    "54.219.41.56:18888",
-    "35.169.113.187:18888",
-    "34.214.241.188:18888",
-    "47.254.146.147:18888",
-    "47.254.144.25:18888",
-    "47.91.246.252:18888",
-    "47.91.216.69:18888",
-    "39.106.220.120:18888"
   ]
 
   maxActiveNodes = 30
 
   p2p {
-    version = 50 # 50: testnet; 101: debug
+    version = 9 # 46: testnet; 101: debug
   }
 
 }
@@ -65,17 +53,6 @@ seed.node = {
   #   "ip:port"
   # ]
   ip.list = [
-    "47.254.16.55:18888",
-    "47.254.18.49:18888",
-    "18.188.111.53:18888",
-    "54.219.41.56:18888",
-    "35.169.113.187:18888",
-    "34.214.241.188:18888",
-    "47.254.146.147:18888",
-    "47.254.144.25:18888",
-    "47.91.246.252:18888",
-    "47.91.216.69:18888",
-    "39.106.220.120:18888"
   ]
 }
 
@@ -91,96 +68,70 @@ genesis.block = {
     {
       accountName = "Devaccount"
       accountType = "AssetIssue"
-      address = "a099357684bc659f5166046b56c95a0e99f1265cbd"
-      balance = "10000000000000000"
+      address = "a0cafbff74a6eb34597ed4ea65d92e3a2cf0fa900b"
+      balance = "10000"
     },
+    # the account of foundation.
     {
       accountName = "Zion"
       accountType = "AssetIssue"
-      address = "a0b4750e2cd76e19dca331bf5d089b71c3c2798548"
-      balance = "15000000000000000"
+      address = "a055ddae14564f82d5b94c7a131b5fcfd31ad6515a"
+      balance = "25000000000000000"
     },
+
+    # the account of payment
     {
       accountName = "Sun"
       accountType = "AssetIssue"
-      address = "a025a3aae39b24a257f95769c701e8d6978ebe9fc5"
+      address = "a04948c2e8a756d9437037dcd8c7e0c73d560ca38d"
       balance = "10000000000000000"
     },
+
+    # the account of coin burn
     {
       accountName = "Blackhole"
       accountType = "AssetIssue"
-      address = "a0559ccf55fadffdf814a42aff331de9688c132612"
+      address = "a0548794500882809695a8a687866e76d4271a146a"
       balance = "-9223372036854775808"
     }
   ]
 
   witnesses = [
     {
-      address: a0904fe896536f4bebc64c95326b5054a2c3d27df6
-      url = "http://Mercury.org",
-      voteCount = 105
-    },
-    {
-      address: a0807337f180b62a77576377c1d0c9c24df5c0dd62
-      url = "http://Venus.org",
-      voteCount = 104
-    },
-    {
-      address: a05430a3f089154e9e182ddd6fe136a62321af22a7
-      url = "http://Earth.org",
-      voteCount = 103
-    },
-    {
-      address: a08beaa1a8e2d45367af7bae7c490b9932a4fa4301
-      url = "http://Mars.org",
-      voteCount = 102
-    },
-    {
-      address: a0b070b2b58f4328e293dc9d6012f59c263d3a1df6
-      url = "http://Jupiter.org",
+      # privKey:c4e7abd0b1f2e4ab38c98b98ad8e81184ac4a3fa30b75e8b8e397ce362ae0c26
+      address: a03e489e58ab3fe7e4a6a9fa70c8e8b006c868cf48
+      url = "http://nanfengpo1.org"
       voteCount = 101
     },
     {
-      address: a00a9309758508413039e4bc5a3d113f3ecc55031d
-      url = "http://Saturn.org",
-      voteCount = 100
+      # privKey:6de7aa0610f877a215429c5d359845b97c3385a929da436edb1d26ebee73bde3
+      address: a0a09b16db113869d0a88eea6e91227122486f84a2
+      url = "http://nanfengpo2.org"
+      voteCount = 102
     },
     {
-      address: a06a17a49648a8ad32055c06f60fa14ae46df94cc1
-      url = "http://Uranus.org",
-      voteCount = 99
+      # privKey:fe734333466c7a7078c54091f1df02b0a92c75b864f82bb6b1dce6801afb352c
+      address: a0e1a64afb81cdf5054387517ac7c3a7a55ab966ed
+      url = "http://nanfengpo3.org"
+      voteCount = 103
     },
     {
-      address: a0ec6525979a351a54fa09fea64beb4cce33ffbb7a
-      url = "http://Neptune.org",
-      voteCount = 98
-    },
-    {
-      address: a0fab5fbf6afb681e4e37e9d33bddb7e923d6132e5
-      url = "http://Pluto.org",
-      voteCount = 97
-    },
-    {
-      address: a014eebe4d30a6acb505c8b00b218bdc4733433c68
-      url = "http://Altair.org",
-      voteCount = 96
-    },
-    {
-      address: a04711bf7afbdf44557defbdf4c4e7aa6138c6331f
-      url = "http://AlphaLyrae.org",
-      voteCount = 95
+      # privKey:8595b427940f3235bdfe6657610cd98cbc9e5b90c00cfa51c035d49e7659bcf1
+      address: a0cafbff74a6eb34597ed4ea65d92e3a2cf0fa900b
+      url = "http://nanfengpo4.org"
+      voteCount = 104
     }
   ]
 
-  timestamp = "0" #2017-8-26 12:00:00
+  timestamp = "9" #2017-8-26 12:00:00
 
   parentHash = "0x0000000000000000000000000000000000000000000000000000000000000000"
 }
 
 localwitness = [
-
+  c4e7abd0b1f2e4ab38c98b98ad8e81184ac4a3fa30b75e8b8e397ce362ae0c26
 ]
 
 block = {
-  needSyncCheck = true # first node : false, other : true
+  needSyncCheck = false # first node : false, other : true
 }
